refactor: split MainWindow responsibilities
- Keep MainWindow.xaml.cs focused on shared state and constructor initialization

- Move Win32 interop/constants, settings, lifecycle, hotkeys, capture helpers, logging, and translation logic into dedicated partial class files

- Preserve existing behavior while reducing the size and complexity of MainWindow.xaml.cs

- Verification: dotnet build GameChatTranslator.sln -p:EnableWindowsTargeting=true succeeds with 0 errors and the existing WFAC010 high-DPI manifest warning
